refactor: :art: sighandler.c file added on makefile, demo_main.c function deleted
demo_main function moved to sighandler.c and makefile src variable modified
