+++ demo_main.c
@@ -6,38 +6,25 @@
 /*   By: minjacho <minjacho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/31 12:20:24 by minjacho          #+#    #+#             */
-/*   Updated: 2023/12/31 17:18:54 by minjacho         ###   ########.fr       */
+/*   Updated: 2024/01/01 15:25:07 by minjacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "demo_minishell.h"
 
-void	sig_handler(int signo)
-{
-	// ^C 뜨는거 없애고 새로운 줄로 출력할 필요.
-	if (signo == SIGINT)
-	{
-		printf("\n");
-		rl_on_new_line();
-		rl_replace_line("", 0);
-		rl_redisplay();
-	}
-	if (signo == SIGQUIT)
-	{
-		// do nothing
-	}
-}
-
 int main()
 {
 	static int	recent_exit = 0;
-	const char	*prompt = "minishell$ ";
+	const char	*prompt = "minishell-1.0$ ";
 	char		*line;
+	pid_t		pid;
 
 	signal(SIGINT, sig_handler);
 	signal(SIGQUIT, sig_handler);
 	while (1)
 	{
+		signal(SIGINT, sig_handler);
+		signal(SIGQUIT, sig_handler);
 		line = readline(prompt);
 		if (!line)
 		{
@@ -50,6 +37,31 @@ int main()
 		{
 			printf("exit\n");
 			break ;
+		}
+		if (ft_strncmp(line, "cat", 4) == 0)
+		{
+			signal(SIGINT, sig_fork_handler);
+			signal(SIGQUIT, sig_fork_handler);
+			pid = fork();
+			if (pid == 0)
+				execve("/bin/cat", NULL, NULL);
+			else
+			{
+				waitpid(pid, &recent_exit, 0);
+			}
+		}
+		if (ft_strncmp(line, "sleep", 4) == 0)
+		{
+			signal(SIGINT, sig_fork_handler);
+			signal(SIGQUIT, sig_fork_handler);
+			pid = fork();
+			char	*sleep_arg[2] = {"sleep", "5"};
+			if (pid == 0)
+				execve("/bin/sleep", sleep_arg, NULL);
+			else
+			{
+				waitpid(pid, &recent_exit, 0);
+			}
 		}
 	}
 	rl_clear_history();
